Separate functionality to files, add optimize_defaults, rename files
