+++ pythonFiles/include/blender_vscode/load_addons.py
@@ -1,6 +1,5 @@
 import atexit
 import os
-import subprocess
 import sys
 import traceback
 from pathlib import Path
@@ -11,7 +10,9 @@ import bpy
 from . import AddonInfo
 from .communication import send_dict_as_json
 from .environment import addon_directories, KEEP_ADDON_INSTALLED
-from .utils import is_addon_legacy, addon_has_bl_info
+from .modify_blender import disable_copy_settings_from_previous_version, disable_addon_remove, disable_extension_remove, optimize_defaults
+from .utils_blender import is_addon_legacy, addon_has_bl_info
+from .utils_files import resolve_link
 
 if bpy.app.version >= (4, 2, 0):
     _EXTENSIONS_DEFAULT_DIR = Path(bpy.utils.user_resource("EXTENSIONS", path="user_default"))
@@ -20,97 +21,27 @@ else:
 _ADDONS_DEFAULT_DIR = Path(bpy.utils.user_resource("SCRIPTS", path="addons"))
 
 
-def _fake_poll(*args, **kwargs):
-    return False
-
-
-def add_warning_label(layout: bpy.types.UILayout, path: str):
-    layout.label(text="You can not remove this repo when using VS Code on windows. If might cause data loss")
-    print(path)
-    if path:
-        try:
-            layout.operator("file.external_operation", text="Open in explorer", icon="FILEBROWSER").filepath = str(path)
-        except AttributeError:
-            pass
-        layout.operator("dev.copy_text", text="Copy addon path", icon="COPYDOWN").text = str(path)
-
-
-def path_from_addon(module):
-    """Copied from bpy.types.PREFERENCES_OT_addon_remove.path_from_addon from Blender 4.2.
-    Implementation in Blender 2.80 does not work correctly
-    """
-    import os
-    import addon_utils
-
-    for mod in addon_utils.modules():
-        if mod.__name__ == module:
-            filepath = mod.__file__
-            if os.path.exists(filepath):
-                if os.path.splitext(os.path.basename(filepath))[0] == "__init__":
-                    return os.path.dirname(filepath), True
-                else:
-                    return filepath, False
-    return None, False
-
-
-def disable_addon_uninstallation():
-    """
-    On windows may lead to data loss as blender is treating junctions as directories.
-
-    Soft link are handled correctly since [blender 2.8](https://developer.blender.org/rBe6ba760ce8fda5cf2e18bf26dddeeabdb4021066)
-    """
-    from pathlib import Path
-
-    bpy.types.PREFERENCES_OT_addon_remove.draw = lambda self, _context: add_warning_label(
-        self.layout, Path(path_from_addon(self.module)[0]).parent
-    )
-    bpy.types.PREFERENCES_OT_addon_remove.execute = lambda _self, _context: {"FINISHED"}
-    bpy.types.PREFERENCES_OT_addon_remove.invoke = lambda self, context, _event: context.window_manager.invoke_popup(
-        self, width=500
-    )
-
-    if bpy.app.version < (4, 2, 0):
-        return
-    bpy.types.USERPREF_MT_extensions_active_repo_remove.poll = _fake_poll
-
-    old_draw = bpy.types.USERPREF_MT_extensions_active_repo_remove.draw
-
-    def limited_remove_repo(self, context):
-        nonlocal old_draw
-        extensions = context.preferences.extensions
-        active_repo_index = extensions.active_repo
-        repo = extensions.repos[active_repo_index]
-        if repo.module == "user_default":
-            add_warning_label(
-                self.layout, path=Path(repo.custom_directory if repo.use_custom_directory else repo.directory).parent
-            )
-            return
-        else:
-            return old_draw(self, context)
-
-    bpy.types.USERPREF_MT_extensions_active_repo_remove.draw = limited_remove_repo
-
-
 def setup_addon_links(addons_to_load: List[AddonInfo]) -> Tuple[List[Dict], List[Dict]]:
-    path_mappings: List[Dict] = []
-
     # always make sure addons are in path, important when running fresh blender install
     # do it always to avoid very confusing logic in the loop below
     os.makedirs(_ADDONS_DEFAULT_DIR, exist_ok=True)
     if str(_ADDONS_DEFAULT_DIR) not in sys.path:
         sys.path.append(str(_ADDONS_DEFAULT_DIR))
 
+    optimize_defaults()
     remove_broken_addon_links()
     if bpy.app.version >= (4, 2, 0):
         remove_broken_extension_links()
 
-    load_status: List[Dict] = []
-    # disable bpy.ops.preferences.copy_prev() is not happy with links that are about to be crated
-    bpy.types.PREFERENCES_OT_copy_prev.poll = _fake_poll
+    disable_copy_settings_from_previous_version()
 
     if sys.platform == "win32":
-        disable_addon_uninstallation()
+        disable_addon_remove()
+        if bpy.app.version >= (4, 2, 0):
+            disable_extension_remove()
 
+    path_mappings: List[Dict] = []
+    load_status: List[Dict] = []
     for addon_info in addons_to_load:
         try:
             load_path = _link_addon_or_extension(addon_info)
@@ -176,48 +107,11 @@ def _remove_duplicate_extension_links(addon_info: AddonInfo):
         existing_extension_with_the_same_target = does_extension_link_exist(addon_info.load_dir)
 
 
-def _resolve_link_windows_cmd(path: Path) -> Optional[str]:
-    IO_REPARSE_TAG_MOUNT_POINT = "0xa0000003"
-    JUNCTION_INDICATOR = f"Reparse Tag Value : {IO_REPARSE_TAG_MOUNT_POINT}"
-    try:
-        output = subprocess.check_output(["fsutil", "reparsepoint", "query", str(path)])
-    except subprocess.CalledProcessError:
-        return None
-    output_lines = output.decode().split(os.linesep)
-    if not output_lines[0].startswith(JUNCTION_INDICATOR):
-        return None
-    TARGET_PATH_INDICATOR = "Print Name:            "
-    for line in output_lines:
-        if line.startswith(TARGET_PATH_INDICATOR):
-            possible_target = line[len(TARGET_PATH_INDICATOR) :]
-            if os.path.exists(possible_target):
-                return possible_target
-
-
-def _resolve_link(path: Path) -> Optional[str]:
-    """Return target if is symlink or jucntion"""
-    try:
-        return os.readlink(str(path))
-    except OSError as e:
-        # OSError: [WinError 4390] The file or directory is not a reparse point
-        if e.winerror == 4390:
-            return None
-        else:
-            raise e
-    except ValueError as e:
-        # there are major differences in python windows junction support (3.7.0 and 3.7.9 give different errors)
-        if sys.platform == "win32":
-            return _resolve_link_windows_cmd(path)
-        else:
-            print("Warning: can not resolve link target", e)
-            return None
-
-
 def does_addon_link_exist(development_directory: Path) -> Optional[Path]:
     """Search default addon path and return first path that links to `development_directory`"""
     for file in os.listdir(_ADDONS_DEFAULT_DIR):
         existing_addon_dir = Path(_ADDONS_DEFAULT_DIR, file)
-        target = _resolve_link(existing_addon_dir)
+        target = resolve_link(existing_addon_dir)
         if target:
             windows_being_windows = target.lstrip(r"\\?")
             if Path(windows_being_windows) == Path(development_directory):
@@ -226,7 +120,7 @@ def does_addon_link_exist(development_directory: Path) -> Optional[Path]:
 
 
 def does_extension_link_exist(development_directory: Path) -> Optional[Path]:
-    """Search all available extension paths and return path that links to `development_directory"""
+    """Search all available extension paths and return path that links to `development_directory`"""
     for repo in bpy.context.preferences.extensions.repos:
         if not repo.enabled:
             continue
@@ -235,7 +129,7 @@ def does_extension_link_exist(development_directory: Path) -> Optional[Path]:
             continue  # repo dir might not exist
         for file in os.listdir(repo_dir):
             existing_extension_dir = Path(repo_dir, file)
-            target = _resolve_link(existing_extension_dir)
+            target = resolve_link(existing_extension_dir)
             if target:
                 windows_being_windows = target.lstrip(r"\\?")
                 if Path(windows_being_windows) == Path(development_directory):
@@ -248,7 +142,7 @@ def remove_broken_addon_links():
         addon_dir = _ADDONS_DEFAULT_DIR / file
         if not addon_dir.is_dir():
             continue
-        target = _resolve_link(addon_dir)
+        target = resolve_link(addon_dir)
         if target and not os.path.exists(target):
             print("INFO: Removing invalid link:", addon_dir, "->", target)
             os.remove(addon_dir)
@@ -264,7 +158,7 @@ def remove_broken_extension_links():
             continue
         for file in os.listdir(repo_dir):
             existing_extension_dir = repo_dir / file
-            target = _resolve_link(existing_extension_dir)
+            target = resolve_link(existing_extension_dir)
             if target and not os.path.exists(target):
                 print("INFO: Removing invalid link:", existing_extension_dir, "->", target)
                 os.remove(existing_extension_dir)
@@ -294,13 +188,7 @@ def load(addons_to_load: List[AddonInfo]):
 
 def make_temporary_link(directory: Union[str, os.PathLike], link_path: Union[str, os.PathLike]):
     if os.path.exists(link_path):
-        try:
-            os.remove(link_path)
-        except PermissionError as ex:
-            print(
-                f'ERROR: Could not remove path "{link_path}" due to insufficient permission. Please remove it manually.'
-            )
-            raise ex
+        os.remove(link_path)
 
     if sys.platform == "win32":
         import _winapi
@@ -309,9 +197,10 @@ def make_temporary_link(directory: Union[str, os.PathLike], link_path: Union[str
     else:
         os.symlink(str(directory), str(link_path), target_is_directory=True)
 
+    if KEEP_ADDON_INSTALLED:
+        return
+
     def cleanup():
-        if KEEP_ADDON_INSTALLED:
-            return
         if not os.path.exists(link_path):
             return
         try:
